Update main.c
Fix

vrolandAPP  6:02 PM
Nice catch, that's a classic overflow right here! esp_timer_get_time() returns an int64_t, but t_poweron is uint32_t. So, after 2^32/10^6/60 = 71.5 min = 1 hour 11 minutes t_poweron overflows and time_since_poweron_m will be > 10. Changing the time of t_poweron to int64_t should fix it :)

+++ examples/terminal/main/main.c
@@ -45,9 +45,9 @@ int max(int a, int b) {
   return a > b ? a : b;
 }
 
-void delay(uint32_t millis) { vTaskDelay(millis / portTICK_PERIOD_MS); }
+void delay(uint64_t millis) { vTaskDelay(millis / portTICK_PERIOD_MS); }
 
-uint32_t millis() { return esp_timer_get_time() / 1000; }
+uint64_t millis() { return esp_timer_get_time() / 1000; }
 
 int log_to_uart(const char* fmt, va_list args) {
     char buffer[256];
